+++ apps/freelance/.env.local
@@ -1,0 +1,2 @@
+NX_APP_API_URL=http://localhost:8000
+NX_APP_GOOGLE_KEY=710744766443-o5mbkprhmhmbd11uftf4e3mh6stk2cgh.apps.googleusercontent.com
